CONFLICT (content): Merge conflict in llvm/test/Transforms/SimplifyCFG/invoke_unwind.ll
  CONFLICT (content): Merge conflict in llvm/lib/Transforms/Scalar/JumpThreading.cpp

+++ llvm/lib/Transforms/Utils/CodeExtractor.cpp
@@ -69,6 +69,7 @@
 #include <iterator>
 #include <map>
 #include <set>
+#include <stdint.h>
 #include <utility>
 #include <vector>
 
@@ -1490,7 +1491,7 @@ void CodeExtractor::calculateNewCallTerminatorWeights(
 
   // Update the branch weights for the exit block.
   Instruction *TI = CodeReplacer->getTerminator();
-  SmallVector<unsigned, 8> BranchWeights(TI->getNumSuccessors(), 0);
+  SmallVector<uint64_t, 8> BranchWeights(TI->getNumSuccessors(), 0);
 
   // Block Frequency distribution with dummy node.
   Distribution BranchDist;
